+++ CesiumJsonReader/include/CesiumJsonReader/JsonReaderOptions.h
@@ -42,8 +42,31 @@ enum class ExtensionState {
   Disabled
 };
 
-class CESIUMJSONREADER_API ExtensionReaderContext {
+/**
+ * @brief Holds options for reading statically-typed data structures from JSON.
+ */
+class CESIUMJSONREADER_API JsonReaderOptions {
 public:
+  /**
+   * @brief Gets a value indicating whether the values of unknown properties are
+   * captured in the {@link ExtensibleObject::unknownProperties} field.
+   *
+   * If this is false, unknown properties are completely ignored.
+   */
+  bool getCaptureUnknownProperties() const {
+    return this->_captureUnknownProperties;
+  }
+
+  /**
+   * @brief Sets a value indicating whether the values of unknown properties are
+   * captured in the {@link ExtensibleObject::unknownProperties} field.
+   *
+   * If this is false, unknown properties are completely ignored.
+   */
+  void setCaptureUnknownProperties(bool value) {
+    this->_captureUnknownProperties = value;
+  }
+
   /**
    * @brief Registers an extension for an object.
    *
@@ -58,7 +81,7 @@ public:
         this->_extensions.emplace(extensionName, ObjectTypeToHandler()).first;
     it->second.insert_or_assign(
         TExtended::TypeName,
-        ExtensionReaderFactory([](const ExtensionReaderContext& context) {
+        ExtensionReaderFactory([](const JsonReaderOptions& context) {
           return std::make_unique<TExtensionHandler>(context);
         }));
   }
@@ -80,7 +103,7 @@ public:
             .first;
     it->second.insert_or_assign(
         TExtended::TypeName,
-        ExtensionHandlerFactory([](const ExtensionReaderContext& context) {
+        ExtensionHandlerFactory([](const JsonReaderOptions& context) {
           return std::make_unique<TExtensionHandler>(context);
         }));
   }
@@ -113,12 +136,13 @@ public:
 private:
   using ExtensionHandlerFactory =
       std::function<std::unique_ptr<IExtensionJsonHandler>(
-          const ExtensionReaderContext&)>;
+          const JsonReaderOptions&)>;
   using ObjectTypeToHandler = std::map<std::string, ExtensionHandlerFactory>;
   using ExtensionNameMap = std::map<std::string, ObjectTypeToHandler>;
 
   ExtensionNameMap _extensions;
   std::unordered_map<std::string, ExtensionState> _extensionStates;
+  bool _captureUnknownProperties = true;
 };
 
 } // namespace CesiumJsonReader